accel/amdxdna: move populate_range() to common code
aie2_populate_range() is not aie2-specific — it operates on generic
umap/HMM infrastructure shared across generations. Move it to aie.c
as amdxdna_populate_range() so that aie4's cmd submit path can reuse
it without duplicating the logic.

(cherry picked from commit b68b1b4142f3f7d939d5c409341308515e6df629)
